# skills/bkd/references/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "BKD API",
    "description": "Kanban board for managing AI coding agents. Issues are assigned to CLI-based AI engines (Claude Code, Codex, Gemini CLI) that execute autonomously.",
    "version": "dev",
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Default"
    }
  ],
...
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ]
          },
          "error": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "Issue": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "projectId": {
            "type": "string"
          },
          "statusId": {
            "type": "string",
            "enum": [
              "todo",
              "working",
              "review",
              "done"
            ]
          },
          "issueNumber": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "sortOrder": {
            "type": "string"
          },
          "useWorktree": {
            "type": "boolean"
          },
          "isPinned": {
            "type": "boolean"
          },
          "keepAlive": {
            "type": "boolean"
          },
          "engineType": {
            "type": [
              "string",
              "null"
            ]
          },
          "sessionStatus": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "pending",
              "running",
              "completed",
              "failed",
              "cancelled",
              null
            ]
          },
          "prompt": {
            "type": [
              "string",
              "null"
            ]
          },
          "externalSessionId": {
            "type": [
              "string",
              "null"
            ]
          },
          "model": {
            "type": [
              "string",
              "null"
            ]
          },
          "statusUpdatedAt": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "projectId",
          "statusId",
          "issueNumber",
          "title",
          "tags",
          "sortOrder",
          "useWorktree",
          "isPinned",
          "keepAlive",
          "engineType",
          "sessionStatus",
          "prompt",
          "externalSessionId",
          "model",
          "statusUpdatedAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateIssue": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 50
            },
            "maxItems": 10
          },
          "statusId": {
            "type": "string",
            "enum": [
              "todo",
              "working",
              "review",
              "done"
            ]
          },
          "useWorktree": {
            "type": "boolean"
          },
          "keepAlive": {
            "type": "boolean"
          },
          "engineType": {
            "type": "string",
            "pattern": "^(claude-code|codex|acp(:.+)?)$",
            "description": "claude-code | codex | acp | acp:<agent>:<model>"
          },
          "model": {
            "type": "string",
            "pattern": "^[\\w./:\\-[\\]]{1,160}$"
          },
          "permissionMode": {
            "type": "string",
            "enum": [
              "auto",
              "supervised",
              "plan"
            ]
          }
        },
        "required": [
          "title",
          "statusId"
        ]
      },
...
      "UpdateIssue": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "minLength": 1,
            "maxLength": 500
          },
          "tags": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string",
              "maxLength": 50
            },
            "maxItems": 10
          },
          "statusId": {
            "type": "string",
            "enum": [
              "todo",
              "working",
              "review",
              "done"
            ]
          },
          "sortOrder": {
            "type": "string",
            "minLength": 1,
            "maxLength": 50,
            "pattern": "^[a-z0-9]+$/i"
          },
          "isPinned": {
            "type": "boolean"
          },
          "keepAlive": {
            "type": "boolean"
          }
        }
      },
...
    },
...
  },
  "paths": {
...
    "/api/projects/{projectId}/issues": {
      "get": {
        "tags": [
          "Issues"
        ],
        "summary": "List issues in project",
        "operationId": "listIssues",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "projectId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Issue list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Issue"
                      }
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Issues"
        ],
        "summary": "Create issue",
        "operationId": "createIssue",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "projectId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateIssue"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created issue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Issue"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "202": {
            "description": "Created and executing",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Issue"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Project not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{projectId}/issues/{issueId}": {
      "get": {
        "tags": [
          "Issues"
        ],
        "summary": "Get issue",
        "operationId": "getIssue",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "projectId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "issueId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Issue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Issue"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Issue not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Issues"
        ],
        "summary": "Update issue",
        "operationId": "updateIssue",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "projectId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "issueId",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateIssue"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated issue",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "$ref": "#/components/schemas/Issue"
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Issue not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Issues"
        ],
        "summary": "Soft-delete issue",
        "operationId": "deleteIssue",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "projectId",
            "in": "path"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "issueId",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id"
                      ]
                    }
                  },
                  "required": [
                    "success",
                    "data"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Issue not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}